requirements.txt
fastapi
uvicorn[standard]
tweepy
apscheduler
pydantic
python-dotenv
pillow
jinja2
google-generativeai
